1)удалил DatabaseConfig и создал yaml с конфигурацией подключения к бд: 2) переименовал эндпоинт для смены пароля 3) создал глобальный класс исключений и убрал логику обработки исключений из контроллеров 4) удалил неиспользуемые дто по фронту: 1) добавил отписку в this.authService.changePassword() с помощью takeUntil 2)убрал лишнюю переменную  ChangePasswordComponent  success

+++ task5/task5/src/main/java/com/example/service/UserService.java
@@ -1,6 +1,9 @@
 package com.example.service;
 
 import com.example.dao.UserDao;
+import com.example.exeption.DatabaseException;
+import com.example.exeption.UserAlreadyExistsException;
+import com.example.exeption.UserNotFoundException;
 import com.example.model.Role;
 import com.example.model.User;
 import org.springframework.stereotype.Service;
@@ -22,61 +25,68 @@ public class UserService {
     }
 
 
-    public Collection<User> getAllUsers() throws SQLException {
-        return userDao.getAllUsers();
-    }
-
-    public User getUserByLogin(String login) throws SQLException {
-        return userDao.getUserByLogin(login);
-    }
-
-    public boolean addUser(String login, String password, String email,
-                           String surname, String name, String patronymic,
-                           LocalDate birthday, Set<Role> roles, Map<String, String> errors) {
+    public Collection<User> getAllUsers()  {
         try {
-            if (userDao.getUserByLogin(login) != null) {
-                errors.put("login", "Пользователь с таким логином уже существует");
-                return false;
-            }
-            User user = new User(login, password, email, surname, name, patronymic, birthday, roles);
-            userDao.addUser(user);
-            return true;
+            return userDao.getAllUsers();
         } catch (SQLException e) {
-            e.printStackTrace();
-            return false;
+            throw new DatabaseException("Ошибка базы данных при получении списка пользователей", e);
         }
     }
 
-    public boolean updateUser(String originalLogin, String login, String password, String email,
-                              String surname, String name, String patronymic,
-                              LocalDate birthday, Set<Role> roles, Map<String, String> errors) {
+    public User getUserByLogin(String login)  {
+        try {
+            User user = userDao.getUserByLogin(login);
+            if (user == null) {
+                throw new UserNotFoundException(login);
+            }
+            return user;
+        } catch (SQLException e) {
+            throw new DatabaseException("Ошибка базы данных при получении пользователя по логину", e);
+        }
+    }
+
+    public void addUser(String login, String password, String email,
+                        String surname, String name, String patronymic,
+                        LocalDate birthday, Set<Role> roles) {
+        try {
+            if (userDao.getUserByLogin(login) != null) {
+                throw new UserAlreadyExistsException(login);
+            }
+            User user = new User(login, password, email, surname, name, patronymic, birthday, roles);
+            userDao.addUser(user);
+
+        } catch (SQLException e) {
+            throw new DatabaseException("Ошибка базы данных при добавлении пользователя", e);
+        }
+    }
+
+    public void updateUser(String originalLogin, String login, String password, String email,
+                           String surname, String name, String patronymic,
+                           LocalDate birthday, Set<Role> roles) {
         try {
             if (!originalLogin.equals(login) && userDao.getUserByLogin(login) != null) {
-                errors.put("login", "Пользователь с таким логином уже существует");
-                return false;
+                throw new UserAlreadyExistsException(login);
             }
 
             User user = new User(login, password, email, surname, name, patronymic, birthday, roles);
             userDao.updateUser(originalLogin, user);
-            return true;
+
         } catch (SQLException e) {
-            e.printStackTrace();
-            return false;
+            throw new DatabaseException("Ошибка базы данных при обновлении пользователя", e);
         }
     }
 
-    public boolean deleteUser(String login) {
+    public void deleteUser(String login) {
         try {
+            User user = userDao.getUserByLogin(login);
+            if (user == null) {
+                throw new UserNotFoundException(login);
+            }
             userDao.deleteUser(login);
-            return true;
         } catch (SQLException e) {
-            e.printStackTrace();
-            return false;
+            throw new DatabaseException("Ошибка базы данных при удалении пользователя", e);
         }
     }
-
-
-
 
 
 }
